requirement.txt
beautifulsoup4==4.11.1
certifi==2022.9.24
charset-normalizer==2.1.1
et-xmlfile==1.1.0
idna==3.4
openpyxl==3.0.10
pip==22.2.2
PyRsa==1.0.3
PyYAML==6.0
requests==2.28.1
setuptools==65.4.1
soupsieve==2.3.2.post1
urllib3==1.26.12
